Issues #39 Bulkのインターフェース化
 - [x] インターフェース化
 - [x] メタジェネレーター修正
