+++ typesafe-query-core/src/test/java/com/sample/model/ApUser_.java
@@ -1,6 +1,7 @@
 package com.sample.model;
 
 import com.github.typesafe_query.Bulk;
+import com.github.typesafe_query.DefaultBulk;
 import java.util.Arrays;
 import com.github.typesafe_query.meta.impl.DBTableImpl;
 import com.github.typesafe_query.meta.impl.DateDBColumnImpl;
@@ -33,7 +34,7 @@ public final class ApUser_{
 	private static final ModelDescription _DESC = new ModelDescription(ApUser.class, _FIELDS);
 	private static final ModelHandler<ApUser> model = new ModelHandler<ApUser>(ApUser.class,TABLE,_DESC);
 	private static final Finder<String,ApUser> find = new DefaultFinder<java.lang.String,ApUser>(ApUser.class,TABLE,_DESC);
-	private static final Bulk bulk = new Bulk(TABLE);
+	private static final Bulk bulk = new DefaultBulk(TABLE);
 
 	private ApUser_(){
 	}
