+++ typesafe-query-core/src/test/java/com/sample/model/Role_.java
@@ -1,6 +1,7 @@
 package com.sample.model;
 
 import com.github.typesafe_query.Bulk;
+import com.github.typesafe_query.DefaultBulk;
 import java.util.Arrays;
 import com.github.typesafe_query.meta.impl.DBTableImpl;
 import com.github.typesafe_query.meta.impl.StringDBColumnImpl;
@@ -25,7 +26,7 @@ public final class Role_{
 	private static final ModelDescription _DESC = new ModelDescription(Role.class, _FIELDS);
 	private static final ModelHandler<Role> model = new ModelHandler<Role>(Role.class,TABLE,_DESC);
 	private static final Finder<RolePK,Role> find = new DefaultFinder<com.sample.model.RolePK,Role>(Role.class,TABLE,_DESC);
-	private static final Bulk bulk = new Bulk(TABLE);
+	private static final Bulk bulk = new DefaultBulk(TABLE);
 
 	private Role_(){
 	}
