+++ typesafe-query-core/src/test/java/com/sample/model/Unit_.java
@@ -1,6 +1,7 @@
 package com.sample.model;
 
 import com.github.typesafe_query.Bulk;
+import com.github.typesafe_query.DefaultBulk;
 import java.util.Arrays;
 import com.github.typesafe_query.meta.impl.DBTableImpl;
 import com.github.typesafe_query.meta.impl.StringDBColumnImpl;
@@ -24,7 +25,7 @@ public final class Unit_{
 	private static final ModelDescription _DESC = new ModelDescription(Unit.class, _FIELDS);
 	private static final ModelHandler<Unit> model = new ModelHandler<Unit>(Unit.class,TABLE,_DESC);
 	private static final Finder<String,Unit> find = new DefaultFinder<java.lang.String,Unit>(Unit.class,TABLE,_DESC);
-	private static final Bulk bulk = new Bulk(TABLE);
+	private static final Bulk bulk = new DefaultBulk(TABLE);
 
 	private Unit_(){
 	}
